Добавили поддержку интеграции SDK на iOS клиентах (#19)
* Swift package provided

* Swift package provided [2]

* Network fixes and arm64 ios simulator support

* Properties module fixes

* Manual properties on ios fix

* Manual properties on ios fix [2]

* Manual properties on ios fix [3]

* Manual properties on ios fix [3.5]

* Manual properties on ios fix [3.9]

* Manual properties on ios fix [3.99]

* Manual properties on ios fix [4]

* Manual properties on ios fix [5]

* Manual properties on ios fix [6]

* Manual properties on ios fix [7]

* Manual properties on ios fix [8]

* Ios dependencies fix

* Ios dependencies fix [2]

* Ios dependencies fix [3]

* Ios dependencies fix [4]

* Ios format fix

* Ios format fix [2]

* Ios format fix [3]

* Kmm bridge test

* Kmm bridge test [2]

* Kmm bridge reset

* Kmm bridge reset [1]

* Unnecessary imports removed

* Event sending fixes

* Device type property update

* Clickstream config size of batch type update

* Http client url fix

* 204 error fix

* 204 error fix [2]

* 204 error fix [3]

* 204 error fix [4]

* Event properties merging and sending fix

* Mindbox id property name update

* Event params fixes

* Event params fixes for ios

* Section type set to string

* Url format updated

+++ clickstream/src/androidMain/kotlin/analytics/sdk/clickstream/AnalyticsJobScheduler.kt
@@ -22,7 +22,7 @@ actual class AnalyticsJobScheduler(private val context: Context) {
             .build()
 
         request = PeriodicWorkRequestBuilder<SendToAnalyticsPeriodicTask>(
-                repeatInterval = clickStreamConfig.sendDataPeriodicityInMinutes,
+                repeatInterval = clickStreamConfig.sendDataPeriodicityInMinutes.toLong(),
                 repeatIntervalTimeUnit = TimeUnit.MINUTES
             )
                 .setConstraints(constraints).build()
